rename : AlarmCommand -> AlarmEvent
