+++ match-service/src/main/java/com/ohdelivery/service/match/alarm/infrastructure/messaging/AlarmEventConsumer.java
@@ -4,7 +4,6 @@ import static com.ohdelivery.common.kafka.Topic.CREATED_MATCHING;
 import static com.ohdelivery.service.match.rider.infrastructure.messaging.RiderKafkaConfig.MATCH_GROUP_ID;
 
 import com.ohdelivery.common.kafka.dto.CreateMatchingEvent;
-import com.ohdelivery.service.match.alarm.application.dto.AlarmCommand;
 import com.ohdelivery.service.match.alarm.application.service.AlarmService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -24,7 +23,7 @@ public class AlarmEventConsumer {
       containerFactory = "createMatchingKafkaListenerContainerFactory"
   )
   public void sendAlarm(CreateMatchingEvent event) {
-    AlarmCommand alarmCommand = AlarmCommand.from(event);
-    alarmService.sendAlarm(alarmCommand);
+    AlarmEvent alarmEvent = AlarmEvent.from(event);
+    alarmService.sendAlarm(alarmEvent);
   }
 }
